가게 관련 코드 수정
- 가게 최대 등록 수 추가
- 가게 정보 추가
- 메소드명 변경

+++ src/main/java/com/hoon/foodrocket/controller/RestaurantController.java
@@ -1,6 +1,8 @@
 package com.hoon.foodrocket.controller;
 
+import com.hoon.foodrocket.application.OwnerService;
 import com.hoon.foodrocket.application.RestaurantService;
+import com.hoon.foodrocket.domain.Owner;
 import com.hoon.foodrocket.domain.Restaurant;
 import com.hoon.foodrocket.util.HttpSessionUtil;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,55 +14,66 @@ import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @RestController
+@RequestMapping("/restaurants")
 public class RestaurantController {
+    public static final int MAX_NUMBER_RESTAURANT = 3;
+
     @Autowired
     private RestaurantService restaurantService;
 
-    @GetMapping("/restaurants")
+    @Autowired
+    private OwnerService ownerService;
+
+    @GetMapping
     public List<Restaurant> list() {
         return restaurantService.getRestaurants();
     }
 
-    @GetMapping("/restaurants/{id}")
+    @GetMapping("/{id}")
     public Restaurant detail(@PathVariable("id") Long id) {
         return restaurantService.getRestaurant(id);
     }
 
-    @PostMapping("/restaurants")
+    @PostMapping
     public ResponseEntity<?> create(@RequestBody Restaurant resource, HttpSession session) {
-        String loginUserEmail = HttpSessionUtil.getLoginUserEmail(session);
+        String loginOwnerEmail = HttpSessionUtil.getLoginOwnerEmail(session);
 
-        if (loginUserEmail == null) {
-            throw new IllegalStateException("로그인이 필요합니다.");
+        if (loginOwnerEmail == null) {
+            throw new IllegalStateException("로그인(사장)이 필요합니다.");
         }
 
-        String name = resource.getName();
-        String address = resource.getAddress();
-        String ownerEmail = resource.getOwnerEmail();
+        int count = restaurantService.getNumberOfRestaurants(loginOwnerEmail);
+        if (count >= MAX_NUMBER_RESTAURANT) {
+            throw new IllegalStateException("가게는 최대 3곳까지만 등록할 수 있습니다.");
+        }
 
-        restaurantService.addRestaurant(name, address, ownerEmail);
+        resource.setOwnerEmail(loginOwnerEmail);
+        restaurantService.registerRestaurant(resource);
 
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
-    @PatchMapping("/restaurants/{id}")
+    @PatchMapping("/{id}")
     public ResponseEntity<?> update(@PathVariable("id") Long id, @RequestBody Restaurant resource, HttpSession session) {
-        String loginUserEmail = HttpSessionUtil.getLoginUserEmail(session);
+        String loginOwnerEmail = HttpSessionUtil.getLoginOwnerEmail(session);
 
-        if (loginUserEmail == null) {
-            throw new IllegalStateException("로그인이 필요합니다.");
+        if (loginOwnerEmail == null) {
+            throw new IllegalStateException("로그인(사장)이 필요합니다.");
         }
 
-        String name = resource.getName();
-        String address = resource.getAddress();
-
-        restaurantService.updateRestaurant(id, name, address, loginUserEmail);
+        restaurantService.updateRestaurant(id, resource, loginOwnerEmail);
 
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 
-    @DeleteMapping("/restaurants/{id}")
-    public ResponseEntity<?> delete(@PathVariable("id") Long id) {
+    @DeleteMapping("/{id}")
+    public ResponseEntity<?> delete(@PathVariable("id") Long id, HttpSession session) {
+        String loginOwnerEmail = HttpSessionUtil.getLoginOwnerEmail(session);
+
+        if (loginOwnerEmail == null) {
+            throw new IllegalStateException("로그인(사장)이 필요합니다.");
+        }
+
         restaurantService.deleteRestaurant(id);
 
         return ResponseEntity.status(HttpStatus.OK).build();
